+++ userspace/lib/include/netinet/tcp.h
@@ -1,0 +1,6 @@
+#ifndef SRVROS_POSIX_NETINET_TCP_H
+#define SRVROS_POSIX_NETINET_TCP_H
+
+#define TCP_NODELAY 1
+
+#endif